rsky-wintermute: fix HANDLE_MAX_TRIES type for LEAST() promotion
The UPDATE actor SET handleResolveTries = LEAST(handleResolveTries + 1, $3)
forces $3 to int4 because smallint + int4 literal promotes the LEAST args
to int4. Sending HANDLE_MAX_TRIES (i16) yielded "cannot convert i16 and
int4" on every identity/sync event, which was firing ~5/sec in production
and silently dropping handle-rename indexing. Cast at the 6 PG callsites
keeps the constant i16 (matches the column for the in-Rust loop / tests)
while satisfying tokio-postgres' parameter type inference.

+++ rsky-wintermute/src/indexer/mod.rs
@@ -368,7 +368,7 @@ impl IndexerManager {
                  SET \"indexedAt\" = $2, \
                      \"handleResolveTries\" = LEAST(\"handleResolveTries\" + 1, $3) \
                  WHERE did = $1",
-                &[&did, &timestamp, &HANDLE_MAX_TRIES],
+                &[&did, &timestamp, &(HANDLE_MAX_TRIES as i32)],
             )
             .await?;
         Ok(())
@@ -1006,7 +1006,7 @@ impl IndexerManager {
                                  \"indexedAt\" = $2, \
                                  \"handleResolveTries\" = LEAST(\"handleResolveTries\" + 1, $3) \
                              WHERE did = $1",
-                            &[&did, &timestamp, &HANDLE_MAX_TRIES],
+                            &[&did, &timestamp, &(HANDLE_MAX_TRIES as i32)],
                         )
                         .await?;
                     return Ok(false);
@@ -1061,7 +1061,7 @@ impl IndexerManager {
                        handle = NULL, \
                        \"indexedAt\" = EXCLUDED.\"indexedAt\", \
                        \"handleResolveTries\" = LEAST(actor.\"handleResolveTries\" + 1, $3)",
-                    &[&did, &timestamp, &HANDLE_MAX_TRIES],
+                    &[&did, &timestamp, &(HANDLE_MAX_TRIES as i32)],
                 )
                 .await?;
         }
